Fix issue with the canvas panning when clicking and moving nodes while zoomed in or out
Allow panning with right and middle mouse buttons, or ctrl+alt+left
mouse button for laptops.
Panning is handled internally by the ScrollPane when ScrollPane.pannable
is set to true, as long as child nodes don't consume the mouse events

Scrolling zooms in toward approximately the place where the mouse
hovering on the canvas. This needs more work.